Extract Loop extraction code
